Updated functions that interface with MM2 using new AGE_API macro.
Methods that use this macro will be exported by the DLL,
as well as explicitly marked as "NOINLINE" to ensure proper behavior.

For example, we add AGE_API to all of the members in datParser that
represent a 1:1 function hook.

However, some methods like AddToken are not originally in the game,
therefore we do not add AGE_API to these functions.

There are some exceptions where AGE_API should be added, such as
the various AddValue helper methods in datParser.

Methods marked explicitly as "inline" should not use AGE_API.