Complete rewrite of hook system, beta support removed.
The hooking system is nothing like before, so be sure to check
out the current implementations to get an idea of how to properly
write new hooks.

This is another one of those giant commits that would take me ages
to write a commit summary for, therefore you'll have to take my word
that there's been a lot of work done and yes, it works.

Good riddance to beta support!