Added custom implementation for gfxPipeline::SetRes and gfxApplySettings.
Support for 32-bit is checked in gfxApplySettings.

If no acceptable depths are found, SetRes will use its
default values. 'cdepth' and 'zdepth' can be overridden individually.

Also added 'bpp' parameter to combine 'cdepth' and 'zdepth' together.
This is automatically detected so it shouldn't need to be used, really.

Extra arguments for SetRes are enabled only on the first call,
and ignored on the rest.

Also added a couple of helpers, gfxWindowMove and gfxWindowUpdate.

Could use some improvements on screen clearing when switching screens.

** Not sure if this is related to these changes I've made, but the game
throws out an error upon exiting. Might want to check that out. **