Fixed incorrect behavior of an AGEHook defined in a namespace.
Prior to this fix, a new instance would be defined every time it was
referenced and required a symbol. Yikes!

This required a new HOOK_API macro, for lack of a better name.
It is not required on static members inside of classes.

All hooks will be declared once "mm2.cpp" is compiled. :)