Implemented a potential fix for CD volume not being updated with mixer.
MM1 uses what is called a "MixerCTL" to send direct commands to mixers,
but this was left unused in MM2 for some reason.

I've re-added the functionality of the mixer by cross-referencing MM1's
implementation, and so far it seems to work just fine.

I have no reason to believe this will cause any crashes, so I haven't
added any configuration options to enable/disable this patch.

If you happen to come across a crash related to audio, please report it!