Added slightly more detailed information when a crash occurs in MM2Hook.
I've yet to get symbols loaded, but at least it'll be easier to do so --
all that's left really is parsing the map/pdb file to get symbol info.

When a crash happens in MM2Hook, it'll print the relative address to
the MM2Hook module, so tracking down the source of a crash will now be
slightly less difficult, especially if it's reported by someone who
doesn't know how to work with a debugger.